Lb3

+++ src/main/java/org/example/bom/Motorcycle.java
@@ -6,7 +6,7 @@ public class Motorcycle extends Vehicle {
 
     private final double annualInterestRate = 0.08; // Річна процентна ставка 8%
 
-    public Motorcycle(Long id, int amount, Model model, Color color, double price, int year) {
+    public Motorcycle(Long id, int amount, Model model, Color color, float price, int year) {
         super(id, amount, model, color, price, year);
     }
